Remove maf ray and move it to pytorch
