+++ python/examples/gpt_oss_20b_finetune/simple_train.py
@@ -11,10 +11,11 @@ from ray.train.lightning import RayFSDPStrategy
 
 import michelangelo.uniflow.core as uniflow
 from examples.gpt_oss_20b_finetune.model import create_gpt_model
-from michelangelo.maf.ray.train import create_run_config, create_scaling_config
 from michelangelo.sdk.trainer.torch.pytorch_lightning.lightning_trainer import (
     LightningTrainer,
     LightningTrainerParam,
+    create_run_config,
+    create_scaling_config,
 )
 from michelangelo.sdk.workflow.variables import DatasetVariable
 from michelangelo.uniflow.plugins.ray import RayTask
