--- python/michelangelo/maf/__init__.py
@@ -1,1 +1,0 @@
-# MAF (Michelangelo AI Framework) module
