--- python/michelangelo/maf/ray/__init__.py
@@ -1,1 +1,0 @@
-# MAF Ray module
