+++ python/michelangelo/sdk/trainer/torch/pytorch_lightning/__init__.py
@@ -1,1 +1,1 @@
-# PyTorch Lightning trainer module
+"""PyTorch Lightning trainer module."""
