--- python/tests/michelangelo/maf/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for michelangelo.maf module."""
