--- python/tests/michelangelo/maf/ray/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for michelangelo.maf.ray module."""
